+++ mjolnir/forcefield/local/CosinePotential.cpp
@@ -1,4 +1,4 @@
-#include <mjolnir/potential/local/CosinePotential.hpp>
+#include <mjolnir/forcefield/local/CosinePotential.hpp>
 
 #ifndef MJOLNIR_SEPARATE_BUILD
 #error "MJOLNIR_SEPARATE_BUILD flag is required"
