+++ mjolnir/forcefield/local/DirectionalContactInteraction.cpp
@@ -1,4 +1,4 @@
-#include <mjolnir/interaction/local/DirectionalContactInteraction.hpp>
+#include <mjolnir/forcefield/local/DirectionalContactInteraction.hpp>
 
 #ifndef MJOLNIR_SEPARATE_BUILD
 #error "MJOLNIR_SEPARATE_BUILD flag is required"
