+++ mjolnir/forcefield/local/GoContactAttractivePotential.cpp
@@ -1,4 +1,4 @@
-#include <mjolnir/potential/local/GoContactAttractivePotential.hpp>
+#include <mjolnir/forcefield/local/GoContactAttractivePotential.hpp>
 
 #ifndef MJOLNIR_SEPARATE_BUILD
 #error "MJOLNIR_SEPARATE_BUILD flag is required"
