+++ mjolnir/forcefield/local/GoContactPotential.cpp
@@ -1,4 +1,4 @@
-#include <mjolnir/potential/local/GoContactPotential.hpp>
+#include <mjolnir/forcefield/local/GoContactPotential.hpp>
 
 #ifndef MJOLNIR_SEPARATE_BUILD
 #error "MJOLNIR_SEPARATE_BUILD flag is required"
