+++ mjolnir/forcefield/local/HarmonicPotential.cpp
@@ -1,4 +1,4 @@
-#include <mjolnir/potential/local/HarmonicPotential.hpp>
+#include <mjolnir/forcefield/local/HarmonicPotential.hpp>
 
 #ifndef MJOLNIR_SEPARATE_BUILD
 #error "MJOLNIR_SEPARATE_BUILD flag is required"
