+++ mjolnir/forcefield/local/PeriodicGaussianPotential.cpp
@@ -1,4 +1,4 @@
-#include <mjolnir/potential/local/PeriodicGaussianPotential.hpp>
+#include <mjolnir/forcefield/local/PeriodicGaussianPotential.hpp>
 
 #ifndef MJOLNIR_SEPARATE_BUILD
 #error "MJOLNIR_SEPARATE_BUILD flag is required"
